backend/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
pillow==12.0.0
openai==2.5.0
python-dotenv==1.1.1
requests
transformers>=4.30.0
torch>=2.0.0
pillow>=9.0.0==2.32.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
imagehash==4.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
replicate==0.34.1
